Integração com front-end

+++ app/adms/Views/Situation/ListSitsUsers.php
@@ -4,24 +4,65 @@ if (!defined('G9C8O7N6N5T4I')) {
     header("Location: /");
     die("Erro: Página não encontrada!");
 }
+?>
 
-echo "<h2>Listar Situações</h2>";
+<!-- Inicio do conteudo do administrativo -->
+<div class="wrapper">
+    <div class="row">
+        <div class="top-list">
+            <span class="title-content">Listar Situação</span>
+            <div class="top-list-right">
+                <?php
+                echo "<a href='" . URLADM . "add-sits-users/index' class='btn-success'>Cadastrar</a>";
+                ?>
+            </div>
+        </div>
+        <div class="content-adm-alert">
+            <?php
+            if (isset($_SESSION['msg'])) {
+                echo $_SESSION['msg'];
+                unset($_SESSION['msg']);
+            }
+            ?>
+        </div>
+        <table class="table-list">
+            <thead class="list-head">
+                <tr>
+                    <th class="list-head-content">ID</th>
+                    <th class="list-head-content">Nome</th>
+                    <th class="list-head-content">Ações</th>
+                </tr>
+            </thead>
+            <tbody class="list-body">
+                <?php
+                foreach ($this->data['listSitsUsers'] as $sitUser) {
+                    extract($sitUser);
+                ?>
+                    <tr>
+                        <td class="list-body-content"><?php echo $id; ?></td>
+                        <td class="list-body-content">
+                            <?php echo "<span style='color: $color'>$name</span>"; ?>
+                        </td>
+                        <td class="list-body-content">
+                            <div class="dropdown-action">
+                                <button onclick="actionDropdown(<?php echo $id; ?>)" class="dropdown-btn-action">Ações</button>
+                                <div id="actionDropdown<?php echo $id; ?>" class="dropdown-action-item">
+                                    <?php
+                                    echo "<a href='" . URLADM . "view-sits-users/index/$id'>Visualizar</a>";
+                                    echo "<a href='" . URLADM . "edit-sits-users/index/$id'>Editar</a>";
+                                    echo "<a href='" . URLADM . "delete-sits-users/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")'>Apagar</a>";
+                                    ?>
+                                </div>
+                            </div>
+                        </td>
+                    </tr>
+                <?php
+                }
+                ?>
+            </tbody>
+        </table>
 
-echo "<a href='" . URLADM . "add-sits-users/index'>Cadastrar nova situação</a><br><br>";
-
-if (isset($_SESSION['msg'])) {
-    printf($_SESSION['msg']);
-    unset($_SESSION['msg']);
-}
-
-foreach ($this->data['listSitsUsers'] as $situation) {
-    extract($situation);
-    echo "ID: $id <br>";
-    echo "Nome: <span style='color: $color;'>$name</span><br>";
-    echo "<a href='" . URLADM . "view-sits-users/index/$id'>Visualizar</a><br>";
-    echo "<a href='" . URLADM . "edit-sits-users/index/$id'>Editar</a><br>";
-    echo "<a href='" . URLADM . "delete-sits-users/index/$id'>Apagar</a>";
-    echo "<hr>";
-}
-
-echo $this->data['pagination'];
+        <?php echo $this->data['pagination']; ?>
+    </div>
+</div>
+<!-- Fim do conteudo do administrativo -->
